feat: fix error - minji

+++ HotSource/src/main/java/hotsource/model/subscribe/SubscribeService.java
@@ -7,6 +7,6 @@ public interface SubscribeService {
 	public List selectByUserId(long user_id);
 	public List selectBySellerId(long seller_id);
 
-	public int selectSubCount(int seller_id);
+	public int selectSubCount(long seller_id);
 
 }
